fix: remove MultisigTest1
